@@ -188,40 +228,27 @@ if __name__ == '__main__':
         ),
     }
 
-    setup(
-        name=NAME,
-        version=VERSION,
-        author='Jon Crall',
-        author_email='jon.crall@kitware.com',
-        url='https://gitlab.kitware.com/computer-vision/netharn',
-        description='Train and deploy pytorch models',
-        long_description=parse_description(),
-        long_description_content_type='text/x-rst',
-        packages=find_packages(include='netharn.*'),
-        package_data={
-            'netharn.initializers._nx_ext_v2': ['*.pyx'],
-        },
-        python_requires='>=3.6',
-        license='Apache 2',
-        classifiers=[
-            # List of classifiers available at:
-            # https://pypi.python.org/pypi?%3Aaction=list_classifiers
-            'Development Status :: 4 - Beta',
-            'Intended Audience :: Developers',
-            'Intended Audience :: Science/Research',
-            'Topic :: Scientific/Engineering',
-            'Topic :: Scientific/Engineering :: Artificial Intelligence',
-            'Topic :: Software Development',
-            'Topic :: Software Development :: Libraries :: Python Modules',
-            'Topic :: Utilities',
-            # This should be interpreted as Apache License v2.0
-            'License :: OSI Approved :: Apache Software License',
-            # Supported Python versions
-            'Programming Language :: Python :: 3.6',
-            'Programming Language :: Python :: 3.7',
-            'Programming Language :: Python :: 3.8',
-            'Programming Language :: Python :: 3.9',
-            'Programming Language :: Python :: 3.10',
-        ],
-        **setupkw,
-    )
+    setupkw["name"] = NAME
+    setupkw["version"] = VERSION
+    setupkw["author"] = "Jon Crall"
+    setupkw["author_email"] = "jon.crall@kitware.com"
+    setupkw["url"] = "https://gitlab.kitware.com/computer-vision/netharn"
+    setupkw["description"] = "Train and deploy pytorch models"
+    setupkw["long_description"] = parse_description()
+    setupkw["long_description_content_type"] = "text/x-rst"
+    setupkw["license"] = "Apache 2"
+    setupkw["packages"] = find_packages(".")
+    setupkw["python_requires"] = ">=3.7"
+    setupkw["classifiers"] = [
+        "Development Status :: 4 - Beta",
+        "Intended Audience :: Developers",
+        "Topic :: Software Development :: Libraries :: Python Modules",
+        "Topic :: Utilities",
+        "License :: OSI Approved :: Apache Software License",
+        "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
+    ]
+    setup(**setupkw)
