@@ -158,10 +158,10 @@ class Yolo(nn.Module):
         if anchors is None:
             anchors = np.array([(1.3221, 1.73145), (3.19275, 4.00944),
                                 (5.05587, 8.09892), (9.47112, 4.84053),
-                                (11.2364, 10.0071)], dtype=np.float)
+                                (11.2364, 10.0071)], dtype=float)
             # np.asarray([(1.08, 1.19), (3.42, 4.41), (6.63, 11.38),
             #                       (9.42, 5.11), (16.62, 10.52)],
-            #                      dtype=np.float)
+            #                      dtype=float)
 
         # Parameters
         self.num_classes = num_classes
