@@ -178,7 +178,7 @@ class Yolo2(layers.AnalyticModule):
         if anchors is None:
             anchors = np.array([(1.3221, 1.73145), (3.19275, 4.00944),
                                 (5.05587, 8.09892), (9.47112, 4.84053),
-                                (11.2364, 10.0071)], dtype=np.float)
+                                (11.2364, 10.0071)], dtype=float)
 
         import ndsampler
         classes = ndsampler.CategoryTree.coerce(classes)
